requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google Drive API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Content extraction
PyPDF2==3.0.1
EbookLib==0.18

# LLM integrations
openai==1.12.0
anthropic==0.18.1

# QR codes
qrcode[pil]==7.4.2
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1

# CORS
python-multipart==0.0.6

# Additional utilities
python-dateutil==2.8.2
